requirements.txt
beautifulsoup4==4.12.3
jsonpath_ng==1.6.1
pandas==2.2.2
python-dotenv==1.0.1
Requests==2.32.3
